+++ src-ui/app/components/event-log/event-log-entry/entry-parsers/muted-audio-device.ts
@@ -1,0 +1,22 @@
+import { EventLogEntryParser } from '../event-log-entry-parser';
+import { EventLogMutedAudioDevice, EventLogType } from '../../../../models/event-log-entry';
+
+export class EventLogMutedAudioDeviceEntryParser extends EventLogEntryParser<EventLogMutedAudioDevice> {
+  entryType(): EventLogType {
+    return 'mutedAudioDevice';
+  }
+
+  override headerInfoTitleParams(entry: EventLogMutedAudioDevice): {
+    [p: string]: string;
+  } {
+    return {
+      deviceName: entry.deviceName.driver
+        ? `${entry.deviceName.display} (${entry.deviceName.driver})`
+        : entry.deviceName.display,
+    };
+  }
+
+  override headerInfoSubTitle(entry: EventLogMutedAudioDevice): string {
+    return 'comp.event-log-entry.type.mutedAudioDevice.reason.' + entry.reason;
+  }
+}